this is a big blob of changes, I'm sorry. My connection was down for a couple of days so I couldn't commit more often...
* fix the ftp class in Prague
* enable input events in the SDL console in Berlin
* continue with the refactoring work of PickTraversal (towards a working UnidrawKit)
* lots of changes to make berlin fully std C++ compliant (it now compiles entirely
  with the latest gcc v3 snapshot)

+++ Berlin/src/ScreenManager.cc
@@ -74,7 +74,7 @@ void ScreenManager::repair()
     u.x *= xres;
     l.y *= yres;
     u.y *= yres;
-    _drawable->flush(l.x, l.y, u.x - l.x, u.y - l.y);
+    _drawable->flush(static_cast<long>(l.x), static_cast<long>(l.y), static_cast<long>(u.x - l.x), static_cast<long>(u.y - l.y));
   }
   _emanager->damage(Region_var(_tmpDamage->_this()));
 }
